Integrate both spellings of getSubjectIcon() logic.
